[update] all

+++ XProtocol/XPacketField.cs
@@ -3,7 +3,7 @@
     public class XPacketField
     {
         public byte FieldID { get; set; }
-        public byte FieldSize { get; set; }
+        public ushort FieldSize { get; set; }
         public byte[] Contents { get; set; }
     }
 }
